SerialIface.ino: add a variant protocol using less bytes
Some games, like Another World, are sampling sounds with the OPL2. The
serial port becomes a bottleneck, whatever the buffer size
is. Therefore, the handshake is modified to let the user opt-in for
shorter binary commands (2 bytes) by using "B0F?" instead of "BUF?".

We use a different function to parse commands in fast mode to avoid
too many "if". I don't know if a simple CPU like the ATMega328P can do
effective branch prediction.