SerialIface: reduce protocol to 2-byte commands
Delay are computed on the client-side instead of the server-side. This
is less precise but this makes the protocol less chatty and able to
handle more complex tunes, like the opening of X-Wing. We use a
semaphore to represent the state of the buffer on the Arduino and a
separate thread handles acknowledgments.