MIDI support for Teensy 2.0 and later
This adds MIDI support for Teensy 2.0 and later which alows to use the
OPL2 Audio Board as a MIDI device. The board can now be used for example
with DosBox to play music when MT-32 is selected, or with your favorite
music editing software.

This is the release candidate for 1.4.2 which request #27 to add MIDI
support for Teensy 2.0++.

The following fixes are made:
* Included a TeenstMidi example tu use the OPL2 Audio Board as MIDI
device
* Added MIDI drum instrument patches
* Fixed an error in instrument documentation
* Fixed typos