Teensy MIDI fix
This fixes issues with the reuse of OPL2 channels in the TeensyMidi
sketch. Issues in the function to get a free OPL2 channel would return
the wrong channel causing clicks and pops. This fix eliminates this
issue mostly, though not fully as it also depends on the instrument
being used.

This fix is in reference to #44